+++ server/ServerMessageFactory.h
@@ -16,7 +16,7 @@ public:
 
     static QByteArray createNewClientMessage(const Account *account);
     static QByteArray createLoginReply(bool response, QString typeError, const Account *account, QList<QString>& nameDocuments);
-    static QByteArray createRegistrationUserReply(bool response, double siteIdOfRegisteredUser);    
+    static QByteArray createRegistrationUserReply(bool response, int siteIdOfRegisteredUser);
     static QByteArray createChangeImageReply(bool response);
     static QByteArray createChangePasswordReply(bool response);
     static QByteArray createOpenFileReply(bool response, QString typeError, QString filename, QString uri, CRDT *crdt, QList<Account*>& accounts, QList<Account>& accountsOffline);
